Agent Operator configures remote-write
(cherry picked from commit c4bdf230d527e19f8b77ca5f938b9254ed344f7d)
