+++ pkg/prometheus/operator_test.go
@@ -1,6 +1,10 @@
 package prometheus
 
-import "testing"
+import (
+	"testing"
+
+	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
+)
 
 func TestStatefulSetKeyToPrometheusKey(t *testing.T) {
 	cases := []struct {
@@ -40,7 +44,6 @@ func TestStatefulSetKeyToPrometheusKey(t *testing.T) {
 	}
 }
 
-
 func TestKeyToStatefulSetKey(t *testing.T) {
 	cases := []struct {
 		name     string
@@ -64,5 +67,67 @@ func TestKeyToStatefulSetKey(t *testing.T) {
 		if c.expected != got {
 			t.Fatalf("Expected key %q got %q", c.expected, got)
 		}
+	}
+}
+
+func TestValidateRemoteWriteConfig(t *testing.T) {
+	cases := []struct {
+		name      string
+		spec      monitoringv1.RemoteWriteSpec
+		expectErr bool
+	}{
+		{
+			name: "with_OAuth2",
+			spec: monitoringv1.RemoteWriteSpec{
+				OAuth2: &monitoringv1.OAuth2{},
+			},
+		}, {
+			name: "with_SigV4",
+			spec: monitoringv1.RemoteWriteSpec{
+				Sigv4: &monitoringv1.Sigv4{},
+			},
+		},
+		{
+			name: "with_OAuth2_and_SigV4",
+			spec: monitoringv1.RemoteWriteSpec{
+				OAuth2: &monitoringv1.OAuth2{},
+				Sigv4:  &monitoringv1.Sigv4{},
+			},
+			expectErr: true,
+		}, {
+			name: "with_OAuth2_and_BasicAuth",
+			spec: monitoringv1.RemoteWriteSpec{
+				OAuth2:    &monitoringv1.OAuth2{},
+				BasicAuth: &monitoringv1.BasicAuth{},
+			},
+			expectErr: true,
+		}, {
+			name: "with_BasicAuth_and_SigV4",
+			spec: monitoringv1.RemoteWriteSpec{
+				BasicAuth: &monitoringv1.BasicAuth{},
+				Sigv4:     &monitoringv1.Sigv4{},
+			},
+			expectErr: true,
+		}, {
+			name: "with_BasicAuth_and_SigV4_and_OAuth2",
+			spec: monitoringv1.RemoteWriteSpec{
+				BasicAuth: &monitoringv1.BasicAuth{},
+				Sigv4:     &monitoringv1.Sigv4{},
+				OAuth2:    &monitoringv1.OAuth2{},
+			},
+			expectErr: true,
+		},
+	}
+	for _, c := range cases {
+		test := c
+		t.Run(test.name, func(t *testing.T) {
+			err := ValidateRemoteWriteSpec(test.spec)
+			if err != nil && !test.expectErr {
+				t.Fatalf("unexpected error occurred: %v", err)
+			}
+			if err == nil && test.expectErr {
+				t.Fatalf("expected an error, got nil")
+			}
+		})
 	}
 }
